requirements.txt
httpx
fastmcp 
dotenv
google-adk
google-adk[eval]
beautifulsoup4
google-generativeai